+++ Text.h
@@ -2,8 +2,8 @@
 #define STRING_H
 #define MAX_TEXT_LENGTH 83886080//1024*1024*80
 typedef struct{
-	int offset;
-	int length;
+	long offset;
+	long length;
 }Text;
 
 //void text_init(Text* text);
